Update standards.html.erb
I updated the introductory part of the Standards section.  After the Introdcutory .  The changes there after are too complex, and I didn't want to mess anything up.  

Will continue editing in Powerpoint and ask Mike on May 20 to make a wholesale swap in of the rest of the Standard section - parts (a), (b), (c) -  replacing current text. with what's in PP.  

Pls wait until MOnday to input anything from PP, or coordinate with me.  I'll be working on the PP version.